Move a few functions to a helper class
